1.0.3
绘制文字坐标支持负数

+++ examples/hello-slint/src/main.rs
@@ -45,7 +45,7 @@ fn main() -> Result<()> {
 
     let text = "花.png";
     let text_width = measure_text(text, 25., canvas.font()).width() as u32;
-    canvas.draw_text(text, WHITE, 25., 150-text_width/2, 270);
+    canvas.draw_text(text, WHITE, 25., 150-text_width as i32/2, 270);
     
     canvas.stroke_line((0, 0), (canvas.width() as i32, canvas.height() as i32), RED);
